backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
groq>=0.5.0
Pillow>=10.0.0
pdf2image>=1.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
pydantic>=2.0.0
rapidfuzz>=3.0.0